% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ch04ParameterEstimation.R
\name{asymprior}
\alias{asymprior}
\title{Prior distribution function with guessing parameter}
\usage{
asymprior(c, alp, bet)
}
\arguments{
\item{c}{guessing parmaeter}

\item{alp}{prior to be set}

\item{bet}{prior to be set}
}
\description{
Prior distribution function with guessing parameter
}
