% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ch02TestItemAnalysis.R
\name{Biserial_Correlation}
\alias{Biserial_Correlation}
\title{Biserial Correlation}
\usage{
Biserial_Correlation(i, t)
}
\arguments{
\item{i}{i is a dichotomous-ordinal variable (0/1). x and y can also be the other way around.}

\item{t}{t is a continuous variable. x and y can also be the other way around.}
}
\value{
The biserial correlation coefficient between the two variables.
}
\description{
A biserial correlation is a correlation between dichotomous-ordinal and
continuous variables.
}
