% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch08BNM.R
\name{BitRespPtn}
\alias{BitRespPtn}
\title{Binary pattern maker}
\usage{
BitRespPtn(n)
}
\arguments{
\item{n}{decimal numbers}
}
\value{
binary patterns
}
\description{
Binary pattern maker
}
\details{
if n <- 1, return 0,1
if n <- 2, return 00,01,10,11
and so on.
}
