% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFitModule.R
\name{calcFitIndices}
\alias{calcFitIndices}
\title{calc Fit Indices}
\usage{
calcFitIndices(chi_A, chi_B, df_A, df_B, nobs)
}
\arguments{
\item{chi_A}{chi-squares for this model}

\item{chi_B}{chi-squares for compared model}

\item{df_A}{degrees of freedom for this model}

\item{df_B}{degrees of freedom for compared model}

\item{nobs}{number of observations for Information criteria}
}
\value{
\describe{
\item{NFI}{Normed Fit Index. Lager values closer to 1.0 indicate a better fit.}
\item{RFI}{Relative Fit Index. Lager values closer to 1.0 indicate a better fit.}
\item{IFI}{Incremental Fit Index. Lager values closer to 1.0 indicate a better fit.}
\item{TLI}{Tucker-Lewis Index. Lager values closer to 1.0 indicate a better fit.}
\item{CFI}{Comparative Fit Inderx. Lager values closer to 1.0 indicate a better fit.}
\item{RMSEA}{Root Mean Square Error of Approximation. Smaller values closer to 0.0 indicate a better fit.}
\item{AIC}{Akaike Information Criterion. A lower value indicates a better fit.}
\item{CAIC}{Consistent AIC.A lower value indicates a better fit.}
\item{BIC}{Bayesian Information Criterion. A lower value indicates a better fit.}
}
}
\description{
A general function that returns the model fit indices.
}
