% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ch00dataFormat.R
\name{dataFormat.long}
\alias{dataFormat.long}
\title{dataFormatter for long-type data}
\usage{
dataFormat.long(data, na = NULL, Sid = NULL, Qid = NULL, Resp = NULL, w = NULL)
}
\arguments{
\item{data}{is a data matrix of the type matrix or data.frame.This must
contain at least three columns to identify the student, the item, and
the response. Additionally, it can include a column for the weight of
the items.}

\item{na}{na argument specifies the numbers or characters to be treated as missing values.}

\item{Sid}{Specify the column number containing the studnet ID label vector.}

\item{Qid}{Specify the column number containing the Question label vector.}

\item{Resp}{Specify the column number containing the Response value vector.}

\item{w}{Specify the column number containing the weight vector.}
}
\value{
\describe{
\item{U}{Data matrix. A matrix with rows representing the sample size and columns
representing the number of items, where elements are either 0 or 1. \eqn{u_{ij}=1} indicates
that student i correctly answered item j, while \eqn{u_{ij}=0} means that student i answered
item j incorrectly.}
\item{ID}{The ID label given by the designated column or function."}
\item{ItemLabel}{The item names given by the provided column names or function.}
\item{Z}{Missing indicator matrix.\eqn{z_{uj}=1} indicates that item j is presented to Student i,
while \eqn{z_{ij}=0} indicates item j is NOT presented to Student i.}
\item{w}{item weight vector}
}
}
\description{
A function to reshape long data into a dataset suitable for Exametrika.
}
