% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ch02TestItemAnalysis.R
\name{Dimensionality}
\alias{Dimensionality}
\title{Dimensionality}
\usage{
Dimensionality(U, na = NULL, Z = NULL, w = NULL)
}
\arguments{
\item{U}{U is a data matrix of the type matrix or data.frame.}

\item{na}{na argument specifies the numbers or characters to be treated as missing values.}

\item{Z}{Z is a missing indicator matrix of the type matrix or data.frame}

\item{w}{w is item weight vector}
}
\value{
Returns a list of class c("exametrika", "Dimensionality") containing:
\itemize{
\item Component: Sequence of component numbers
\item Eigenvalue: Eigenvalues of the tetrachoric correlation matrix
\item PerOfVar: Percentage of variance explained by each component
\item CumOfPer: Cumulative percentage of variance explained
}
}
\description{
The dimensionality is the number of components
the test is measuring.
}
