% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ch02TestAnalysis.R
\name{DimensonalityAnalysis}
\alias{DimensonalityAnalysis}
\title{Dimensionality Analysis}
\usage{
DimensonalityAnalysis(U, na = NULL, Z = NULL, w = NULL)
}
\arguments{
\item{U}{U is a data matrix of the type matrix or data.frame.}

\item{na}{na argument specifies the numbers or characters to be treated as missing values.}

\item{Z}{Z is a missing indicator matrix of the type matrix or data.frame}

\item{w}{w is item weight vector}
}
\description{
The dimensionality is the number of contens ( or conponents)
the test is measuring.
}
