% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ch04AbilityEstimation.R
\name{EAP_PSD}
\alias{EAP_PSD}
\title{Ability Parameter Estimation Function}
\usage{
EAP_PSD(Lambda, U, Z)
}
\arguments{
\item{Lambda}{Item parameters matrix}

\item{U}{U is a data matrix of the type matrix or data.frame.}

\item{Z}{Z is a missing indicator matrix of the type matrix or data.frame}
}
\description{
A function that returns the Expected A Posteriori (EAP) of the
ability parameter and its posterior standard deviation.
}
