% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMclus.R
\name{emclus}
\alias{emclus}
\title{emclus funciton}
\usage{
emclus(
  U,
  Z,
  ncls,
  Fil,
  beta1,
  beta2,
  maxiter = 100,
  mic = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{U}{response matrix U of the examData class.}

\item{Z}{missing indicator matrix Z of the examData class.}

\item{ncls}{number of latent class}

\item{Fil}{Filter matrix}

\item{beta1}{beta distribution parameter1 as prior density of rank reference matrix}

\item{beta2}{beta distribution parameter2 as prior density of rank reference matrix}

\item{maxiter}{Maximum number of iterations.}

\item{mic}{Monotonic increasing IRP option}

\item{verbose}{verbose output Flag. default is FALSE}
}
\description{
This function is used for both LCA (Latent Class Analysis) and
LRA (Latent Rank Analysis). LCA is considered a special case of LRA,
in which the filtering matrix is reduced to the identity matrix.
This function takes a dataset, the number of classes, and a filtering
matrix as input and returns the latent rank.
}
