% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch08BNM.R
\name{fill_adj}
\alias{fill_adj}
\title{Data Frame DAG to adj}
\usage{
fill_adj(DAG, ItemLabel)
}
\arguments{
\item{DAG}{A dataframe-type DAG (Directed Acyclic Graph). Descriptions
for all edges are necessary, but ,of course,information on all nodes
is not always required.}

\item{ItemLabel}{Labels for all items. Information should held by the
Exametrika class data.}
}
\value{
adjacency matrix
}
\description{
inner function of return full length adj from data.frame
}
