% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_QitemFunctions.R
\name{generate_category_labels}
\alias{generate_category_labels}
\title{Generate category labels for response data}
\usage{
generate_category_labels(data_column, item_name)
}
\arguments{
\item{data_column}{A vector containing response data for a single item}

\item{item_name}{Character string of the item name}
}
\value{
A character vector of category labels
}
\description{
Generate category labels for response data
}
\details{
If the input is a factor, returns its levels.
Otherwise, generates labels in the format "Item name-Category-N"
}
\keyword{internal}
