% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_GRM.R
\name{generate_start_values}
\alias{generate_start_values}
\title{generate start values for optimize}
\usage{
generate_start_values(tmp)
}
\arguments{
\item{tmp}{dataset}
}
\description{
generate start values for optimize
}
\keyword{internal}
