% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_grm.R
\name{GRM}
\alias{GRM}
\title{Graded Response Model}
\usage{
GRM(U, na = NULL, Z = NULL, w = NULL, verbose = TRUE)
}
\arguments{
\item{U}{U is either a data class of exametrika, or raw data. When raw data is given,
it is converted to the exametrika class with the \code{\link{dataFormat}} function.}

\item{na}{na argument specifies the numbers or characters to be treated as missing values.}

\item{Z}{Z is a missing indicator matrix of the type matrix or data.frame}

\item{w}{w is item weight vector}

\item{verbose}{logical; if TRUE, shows progress of iterations (default: TRUE)}
}
\value{
\describe{
\item{testlength}{Length of the test. The number of items included in the test.}
\item{nobs}{Sample size. The number of rows in the dataset.}
\item{params}{Matrix containing the estimated item parameters}
\item{itemPSD}{Posterior standard deviation of the item parameters}
\item{ability}{Estimated parameters of students ability}
\item{ItemFitIndices}{Fit index for each item.See also \code{\link{ItemFit}}}
\item{TestFitIndices}{Overall fit index for the test.See also \code{\link{TestFit}}}
}
}
\description{
A short description...
}
\examples{
\donttest{
result <- GRM(J5S1000)
print(result)
plot(result, type = "IRF")
plot(result, type = "IIF")
plot(result, type = "TIF")
}
}
