% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_GRM.R
\name{grm_cumprob}
\alias{grm_cumprob}
\title{cumulative probability of GRM}
\usage{
grm_cumprob(theta, a, b)
}
\arguments{
\item{theta}{latent score of subject}

\item{a}{discriminant parameter of IRF}

\item{b}{difficulty parameter of IRF}
}
\description{
cumulative probability of GRM
}
\keyword{internal}
