% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_grm.R
\name{grm_prob}
\alias{grm_prob}
\title{probability function of GRM}
\usage{
grm_prob(theta, a, b)
}
\arguments{
\item{theta}{latent score of subject}

\item{a}{discriminant parameter of IRF}

\item{b}{difficulty parameter of IRF}
}
\description{
probability function of GRM
}
