% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ch04IRTbaseFunctions.R
\name{IIF2PLM}
\alias{IIF2PLM}
\title{IIF for 2PLM}
\usage{
IIF2PLM(a, b, theta)
}
\arguments{
\item{a}{slope parameter}

\item{b}{location parameter}

\item{theta}{ability parameter}
}
\value{
Returns a numeric vector representing the item information at each ability
level theta. The information is calculated as:
\eqn{I(\theta) = a^2P(\theta)(1-P(\theta))}
}
\description{
Item Information Function for 2PLM
}
