% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ch04IRTbaseFunctions.R
\name{IIF3PLM}
\alias{IIF3PLM}
\title{IIF for 3PLM}
\usage{
IIF3PLM(a, b, c, theta)
}
\arguments{
\item{a}{slope parameter}

\item{b}{location parameter}

\item{c}{lower asymptote parameter}

\item{theta}{ability parameter}
}
\value{
Returns a numeric vector representing the item information at each ability
level theta. The information is calculated as:
\eqn{I(\theta) = \frac{a^2(1-P(\theta))(P(\theta)-c)^2}{(1-c)^2P(\theta)}}
}
\description{
Item Information Function for 3PLM
}
