% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_IRM.R
\name{IRM}
\alias{IRM}
\title{IRM (Deprecated)}
\usage{
IRM(...)
}
\arguments{
\item{...}{All arguments passed to \code{\link{Biclustering_IRM}}}
}
\description{
This function has been renamed to \code{\link{Biclustering_IRM}}.
Please use \code{Biclustering_IRM} instead.
}
