% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ch04ParameterEstimation.R
\name{IRT}
\alias{IRT}
\title{Estimating Item parmateres using EM algorithm}
\usage{
IRT(model, U, na = NULL, Z = NULL, w = NULL)
}
\arguments{
\item{model}{This argument takes the number of item parameters to be
estimated in the logistic model. It is limited to values 2, 3, or 4.}

\item{U}{U is a data matrix of the type matrix or data.frame.}

\item{na}{na argument specifies the numbers or characters to be treated as missing values.}

\item{Z}{Z is a missing indicator matrix of the type matrix or data.frame}

\item{w}{w is item weight vector}
}
\description{
A function for estimating item parameters using the EM algorithm.
}
