% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ch04IRTbaseFunctions.R
\name{ItemInformationFunc}
\alias{ItemInformationFunc}
\title{IIF for 4PLM}
\usage{
ItemInformationFunc(a = 1, b, c = 0, d = 1, theta)
}
\arguments{
\item{a}{slope parameter}

\item{b}{location parameter}

\item{c}{lower asymptote parameter}

\item{d}{upper asymptote parameter}

\item{theta}{ability parameter}
}
\value{
Returns a numeric vector representing the item information at each ability
level theta. The information is calculated based on the first derivative of
the log-likelihood of the 4PL model with respect to theta.
}
\description{
Item Information Function for 4PLM
}
