% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_QitemFunctions.R
\name{ItemReport}
\alias{ItemReport}
\title{Item Report for non-binary data}
\usage{
ItemReport(U, na = NULL, Z = NULL, w = NULL)
}
\arguments{
\item{U}{U is a data matrix of the type matrix or data.frame.}

\item{na}{na argument specifies the numbers or characters to be treated as missing values.}

\item{Z}{Z is a missing indicator matrix of the type matrix or data.frame}

\item{w}{w is item weight vector}
}
\value{
\item{Obs}{Number of valid responses for each item}
\item{ObsRatio}{Proportion of valid responses for each item (range: 0-1)}
\item{ItemMean}{Mean score of each item}
\item{ItemSD}{Standard deviation of each item score}
\item{ItemCORR}{Item-total correlation coefficients.
Correlation between item scores and total test scores}
\item{ItemCORR_R}{Corrected item-total correlation coefficients.
Correlation between item scores and total test scores excluding the target item}
}
\description{
Item Report for non-binary data
}
