% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_dataset_explain.R
\docType{data}
\name{J12S5000}
\alias{J12S5000}
\title{J12S5000}
\format{
An exametrika class object with 5000 students and 12 items containing binary (0/1) responses
}
\source{
\url{http://shojima.starfree.jp/tde/index.htm}
}
\usage{
J12S5000
}
\description{
A binary response dataset for test analysis
}
\keyword{datasets}
