% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_dataset_explain.R
\docType{data}
\name{J15S3810}
\alias{J15S3810}
\title{J15S3810}
\format{
An exametrika class object with 3810 students and 15 items containing nominal responses with 4 categories
}
\usage{
J15S3810
}
\description{
A ordinal response dataset for test analysis
}
\keyword{datasets}
