% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_dataset_explain.R
\docType{data}
\name{J15S500}
\alias{J15S500}
\title{J15S500}
\format{
An exametrika class object with 500 students and 15 items containing binary (0/1) responses
}
\source{
\url{http://shojima.starfree.jp/tde/index.htm}
}
\usage{
J15S500
}
\description{
A binary response dataset for test analysis
}
\keyword{datasets}
