% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_dataset_explain.R
\docType{data}
\name{J20S400}
\alias{J20S400}
\title{J20S400}
\format{
An exametrika class object with 400 students and 20 items containing binary (0/1) responses
}
\source{
\url{http://shojima.starfree.jp/tde/index.htm}
}
\usage{
J20S400
}
\description{
A binary response dataset for test analysis
}
\keyword{datasets}
