% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_dataset_explain.R
\docType{data}
\name{J35S5000}
\alias{J35S5000}
\title{J35S5000}
\format{
An exametrika class object with 5000 students and 35 items containing polytomous responses with correct answers
}
\usage{
J35S5000
}
\description{
A rated response dataset for test analysis
}
\keyword{datasets}
