% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_dataset_explain.R
\docType{data}
\name{J35S515}
\alias{J35S515}
\title{J35S515}
\format{
An exametrika class object with 515 students and 35 items containing binary (0/1) responses
}
\source{
\url{http://shojima.starfree.jp/tde/index.htm}
}
\usage{
J35S515
}
\description{
A binary response dataset for test analysis
}
\keyword{datasets}
