% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_dataset_explain.R
\docType{data}
\name{J50S100}
\alias{J50S100}
\title{J50S100}
\format{
An exametrika class object with 100 students and 50 items containing binary responses
}
\usage{
J50S100
}
\description{
A simulated binary dataset for test analysis. This is a synthetic dataset
generated using random number generation for demonstration and testing purposes.
}
\keyword{datasets}
