% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_dataset_explain.R
\docType{data}
\name{J5S10}
\alias{J5S10}
\title{J5S10}
\format{
An exametrika class object with 5 students and 10 items containing binary (0/1) responses
}
\source{
\url{http://sh0j1ma.stars.ne.jp/exmk/}
}
\usage{
J5S10
}
\description{
A binary response dataset for test analysis
}
\keyword{datasets}
