% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_dataset_explain.R
\docType{data}
\name{J5S1000}
\alias{J5S1000}
\title{J5S1000}
\format{
An exametrika class object with 1000 students and 5 items containing ordinal responses
}
\usage{
J5S1000
}
\description{
A simulated ordinal dataset for test analysis. This is a synthetic dataset
generated using random number generation for demonstration and testing purposes.
}
\keyword{datasets}
