% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_GRM.R
\name{Jacobian_grm}
\alias{Jacobian_grm}
\title{Jacobian function for GRM}
\usage{
Jacobian_grm(target, nitems, ncat)
}
\arguments{
\item{target}{target vector}

\item{nitems}{number of items}

\item{ncat}{number of categories for each items}
}
\description{
Jacobian function for GRM
}
\keyword{internal}
