% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_GRM.R
\name{log_lik_grm}
\alias{log_lik_grm}
\title{log_lik function for grm}
\usage{
log_lik_grm(target, dat, verbose)
}
\arguments{
\item{target}{target vector}

\item{dat}{data set}
}
\description{
log_lik function for grm
}
\keyword{internal}
