% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ch04IRTbaseFunctions.R
\name{LogisticModel}
\alias{LogisticModel}
\title{Four-Parameter Logistic Model}
\usage{
LogisticModel(a = 1, b, c = 0, d = 1, theta)
}
\arguments{
\item{a}{slope parameter}

\item{b}{location parameter}

\item{c}{lower asymptote parameter}

\item{d}{upper asymptote parameter}

\item{theta}{ability parameter}
}
\value{
Returns a numeric vector of probabilities between c and d, representing
the probability of a correct response given the ability level theta. The probability
is calculated using the formula: \eqn{P(\theta) = c + \frac{(d-c)}{1 + e^{-a(\theta-b)}}}
}
\description{
The four-parameter logistic model is a model where one additional
parameter d, called the upper asymptote parameter, is added to the
3PLM.
}
