% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFitModule.R
\name{Model_Fit}
\alias{Model_Fit}
\title{Model Fit Functions}
\usage{
Model_Fit(ell_A, ell_B, ell_N, df_A, df_B, nobs)
}
\arguments{
\item{ell_A}{log likelihood of this model}

\item{ell_B}{log likelihood of saturated model}

\item{ell_N}{log likelihood of Null model}

\item{df_A}{degrees of freedom of this model}

\item{df_B}{degrees of freedom of saturated model}

\item{nobs}{number of observations}
}
\description{
A general function that returns the model fit indices.
}
