% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_GRM.R
\name{params_to_target}
\alias{params_to_target}
\title{convert parameters to optimization target}
\usage{
params_to_target(a_vec, b_list)
}
\arguments{
\item{a_vec}{discrimination parameter vector}

\item{b_list}{threshold parameter list}
}
\description{
convert parameters to optimization target
}
\keyword{internal}
