% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_GRM.R
\name{params_to_target_jac}
\alias{params_to_target_jac}
\title{parameter transformation params_to_target}
\usage{
params_to_target_jac(a_vec, b_list)
}
\arguments{
\item{a_vec}{vector of descriminant parameters}

\item{b_list}{lists of difficuluty parameter vec}
}
\description{
parameter transformation params_to_target
}
\keyword{internal}
