% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_TestItemFunctions.R
\name{PolychoricCorrelationMatrix}
\alias{PolychoricCorrelationMatrix}
\title{Polychoric Correlation Matrix}
\usage{
PolychoricCorrelationMatrix(U, na = NULL, Z = NULL, w = NULL)
}
\arguments{
\item{U}{U is a data matrix of the type matrix or data.frame.}

\item{na}{na argument specifies the numbers or characters to be treated as missing values.}

\item{Z}{Z is a missing indicator matrix of the type matrix or data.frame}

\item{w}{w is item weight vector}
}
\value{
A matrix of tetrachoric correlations with exametrika class.
Each element (i,j) represents the tetrachoric correlation between items i and j.
The matrix is symmetric with ones on the diagonal.
}
\description{
Polychoric Correlation Matrix
}
\examples{
\donttest{
# example code
PolychoricCorrelationMatrix(J5S1000)
}
}
