% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ch04ParameterEstimation.R
\name{PSD_item_params}
\alias{PSD_item_params}
\title{internal functions for PSD of Item parameters}
\usage{
PSD_item_params(model, Lambda, quadrature, marginal_posttheta)
}
\arguments{
\item{model}{2,3,or 4PL}

\item{Lambda}{item paramters Matrix}

\item{quadrature}{quads}

\item{marginal_posttheta}{marginal post theta}
}
\description{
internal functions for PSD of Item parameters
}
