% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04A_IRTbaseFunctions.R
\name{RaschModel}
\alias{RaschModel}
\title{Rasch Model}
\usage{
RaschModel(b, theta)
}
\arguments{
\item{b}{slope parameter}

\item{theta}{ability parameter}
}
\value{
Returns a numeric vector of probabilities between 0 and 1, representing
the probability of a correct response given the ability level theta. The probability
is calculated using the formula: \eqn{P(\theta) = \frac{1}{1 + e^{-(\theta-b)}}}
}
\description{
The one-parameter logistic model is a model with only one parameter b.
This model is a 2PLM model in which a is constrained to 1.
This model is also called the Rasch model.
}
