% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_typeErrorMessage.R
\name{response_type_error}
\alias{response_type_error}
\title{Generate Error Message for Invalid Response Type}
\usage{
response_type_error(response_type, fun_name)
}
\arguments{
\item{response_type}{character. One of "binary", "rated", "ordinal", or "nominal"}

\item{fun_name}{character. Name of the calling function}
}
\value{
Never returns; always stops with an error message
}
\description{
Internal function to generate standardized error messages when a function
is called with an incompatible response type.
}
\keyword{internal}
