% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_GRM.R
\name{score_function_with_Jacobian}
\alias{score_function_with_Jacobian}
\title{score function for grm}
\usage{
score_function_with_Jacobian(target, dat)
}
\arguments{
\item{target}{target vector}

\item{dat}{data set}
}
\description{
score function for grm
}
\keyword{internal}
