% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_QitemFunctions.R
\name{ScoreReport}
\alias{ScoreReport}
\title{Score Report for non-binary data}
\usage{
ScoreReport(U, na = NULL, Z = NULL, w = NULL)
}
\arguments{
\item{U}{U is a data matrix of the type matrix or data.frame.}

\item{na}{na argument specifies the numbers or characters to be treated as missing values.}

\item{Z}{Z is a missing indicator matrix of the type matrix or data.frame}

\item{w}{w is item weight vector}
}
\value{
\item{TestLength}{Length of the test. The number of items included in the test.}
\item{SampleSize}{Sample size. The number of rows in the dataset.}
\item{Mean}{Average number of correct answers.}
\item{SEofMean}{Standard error of mean}
\item{Variance}{Variance}
\item{SD}{Standard Deviation}
\item{Skewness}{Skewness}
\item{Kurtosis}{Kurtosis}
\item{Min}{Minimum score}
\item{Max}{Max score}
\item{Range}{Range of score}
\item{Alpha}{Cronbach's alpha coefficient, a measure of internal consistency reliability. }
}
\description{
Score Report for non-binary data
}
