% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08C_BNM_GA.R
\name{StrLearningGA_BNM}
\alias{StrLearningGA_BNM}
\title{StrLearningGA_BNM (Deprecated)}
\usage{
StrLearningGA_BNM(...)
}
\arguments{
\item{...}{All arguments passed to \code{\link{BNM_GA}}}
}
\description{
This function has been renamed to \code{\link{BNM_GA}}.
Please use \code{BNM_GA} instead.
}
