% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08C_BNM_GA.R
\name{StrLearningPBIL_BNM}
\alias{StrLearningPBIL_BNM}
\title{StrLearningPBIL_BNM (Deprecated)}
\usage{
StrLearningPBIL_BNM(...)
}
\arguments{
\item{...}{All arguments passed to \code{\link{BNM_PBIL}}}
}
\description{
This function has been renamed to \code{\link{BNM_PBIL}}.
Please use \code{BNM_PBIL} instead.
}
