% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09B_LDLRA_GA.R
\name{StrLearningPBIL_LDLRA}
\alias{StrLearningPBIL_LDLRA}
\title{StrLearningPBIL_LDLRA (Deprecated)}
\usage{
StrLearningPBIL_LDLRA(...)
}
\arguments{
\item{...}{All arguments passed to \code{\link{LDLRA_PBIL}}}
}
\description{
This function has been renamed to \code{\link{LDLRA_PBIL}}.
Please use \code{LDLRA_PBIL} instead.
}
