% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_GRM.R
\name{target_to_params_jac}
\alias{target_to_params_jac}
\title{parameter transformation target_to_params}
\usage{
target_to_params_jac(target, nitems, ncat)
}
\arguments{
\item{target}{optimize target vector}

\item{nitems}{number of items}

\item{ncat}{number of categories for each items}
}
\description{
parameter transformation target_to_params
}
\keyword{interenal}
