% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ch04IRTbaseFunctions.R
\name{TestInformationFunc}
\alias{TestInformationFunc}
\title{TIF for IRT}
\usage{
TestInformationFunc(params, theta)
}
\arguments{
\item{params}{parameter matrix}

\item{theta}{ability parameter}
}
\value{
Returns a numeric vector representing the test information at each ability
level theta. The test information is the sum of item information functions for
all items in the test: \eqn{I_{test}(\theta) = \sum_{j=1}^n I_j(\theta)}
}
\description{
Test Information Function for 4PLM
}
