% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04A_IRTbaseFunctions.R
\name{ThreePLM}
\alias{ThreePLM}
\title{Three-Parameter Logistic Model}
\usage{
ThreePLM(a, b, c, theta)
}
\arguments{
\item{a}{slope parameter}

\item{b}{location parameter}

\item{c}{lower asymptote parameter}

\item{theta}{ability parameter}
}
\value{
Returns a numeric vector of probabilities between c and 1, representing
the probability of a correct response given the ability level theta. The probability
is calculated using the formula: \eqn{P(\theta) = c + \frac{1-c}{1 + e^{-a(\theta-b)}}}
}
\description{
The three-parameter logistic model is a model where the lower
asymptote parameter c is added to the 2PLM
}
