% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04A_IRTbaseFunctions.R
\name{TwoPLM}
\alias{TwoPLM}
\title{Two-Parameter Logistic Model}
\usage{
TwoPLM(a, b, theta)
}
\arguments{
\item{a}{slope parameter}

\item{b}{location parameter}

\item{theta}{ability parameter}
}
\value{
Returns a numeric vector of probabilities between 0 and 1, representing
the probability of a correct response given the ability level theta. The probability
is calculated using the formula: \eqn{P(\theta) = \frac{1}{1 + e^{-a(\theta-b)}}}
}
\description{
The two-parameter logistic model is a classic model that defines
the probability of a student with ability theta successfully
answering item j, using both a slope parameter and a
location parameter.
}
