% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch11BINET.R
\name{BINET}
\alias{BINET}
\title{Bicluster Network Model}
\usage{
BINET(
  U,
  Z = NULL,
  w = NULL,
  na = NULL,
  conf = NULL,
  ncls = NULL,
  nfld = NULL,
  g_list = NULL,
  adj_list = NULL,
  adj_file = NULL,
  verbose = FALSE
)
}
\arguments{
\item{U}{U is either a data class of Exametrika, or raw data. When raw data is given,
it is converted to the Exametrika class with the \link{dataFormat} function.}

\item{Z}{Z is a missing indicator matrix of the type matrix or data.frame}

\item{w}{w is item weight vector}

\item{na}{na argument specifies the numbers or characters to be treated as missing values.}

\item{conf}{For the confirmatory parameter, you can input either a vector with
items and corresponding fields in sequence, or a field membership profile
matrix. In the case of the former, the field membership profile matrix will be generated internally.
When providing a membership profile matrix, it needs to be either matrix or data.frame.
The number of fields(nfld) will be overwrite to the number of columns of this matrix.}

\item{ncls}{number of classes}

\item{nfld}{number of fields}

\item{g_list}{A list compiling graph-type objects for each rank/class.}

\item{adj_list}{A list compiling matrix-type adjacency matrices for each rank/class.}

\item{adj_file}{A file detailing the relationships of the graph for each rank/class,
listed in the order of starting point, ending point, and rank(class).}

\item{verbose}{verbose output Flag. default is TRUE}
}
\value{
\describe{
\item{nobs}{Sample size. The number of rows in the dataset.}
\item{testlength}{Length of the test. The number of items included in the test.}
\item{Nclass}{Optimal number of classes.}
\item{Nfield}{Optimal number of fields.}
\item{crr}{Correct Response Rate}
\item{ItemLabel}{Label of Items}
\item{FieldLabel}{Label of Fields}
\item{all_adj}{Integrated Adjacency matrix used to plot graph.}
\item{all_g}{Integrated graph object used to plot graph.see also
\link{plot.Exametrika}}
\item{adj_list}{List of Adjacency matrix used in the model}
\item{params}{A list of the estimated conditional probabilities.
It indicates which path was obtained from which parent node(class) to
which child node(class), held by \code{parent}, \code{child}, and \code{field}. The item
Items contained in the field is in \code{fld}. Named \code{chap} includes the
conditonal correct response answer rate of the child node, while \code{pap}
contains the pass rate of the parent node.}
\item{PSRP}{Response pattern by the students belonging to the parent
classes of Class c. A more comprehensible arrangement of \code{params.}}
\item{LCD}{Latent Class Distribution. see also \link{plot.Exametrika}}
\item{LFD}{Latent Field Distribution. see also \link{plot.Exametrika}}
\item{CMD}{Class Membership Distribution.}
\item{FRP}{Marginal bicluster reference matrix.}
\item{FRPIndex}{Index of FFP includes the item location parameters B and Beta,
the slope parameters A and Alpha, and the monotonicity indices C and Gamma.}
\item{TRP}{Test Reference Profile}
\item{LDPSR}{A rearranged set of parameters for output. It includes
the field the items contained within that field, and the conditional
correct response rate of parent nodes(class) and child node(class).}
\item{FieldEstimated}{Given vector which correspondence between items
and the fields.}
\item{Students}{Rank Membership Profile matrix.The s-th row vector of \eqn{\hat{M}_R}, \eqn{\hat{m}_R}, is the
rank membership profile of Student s, namely the posterior probability distribution representing the student's
belonging to the respective latent classes. }
\item{NextStage}{The next class that easiest for students to move to,
its membership probability, class-up odds, and the field required for
mobe.}
\item{MG_FitIndices}{Multigroup as Null model.See also \link{TestFit}}
\item{SM_FitIndices}{Saturated Model as Null model.See also \link{TestFit}}
}
}
\description{
Bicluster Network Model: BINET is a model that combines the Bayesian
network model and Biclustering. BINET is very similar to LDB and LDR.
The most significant difference is that in LDB, the nodes represent
the fields, whereas in BINET, they represent the class. BINET
explores the local dependency structure among latent classes at each
latent field, where each field is a locus.
}
