% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ch02SingleItemAnalysis.R
\name{Biserial_Correlation}
\alias{Biserial_Correlation}
\title{Biserial Correlation}
\usage{
Biserial_Correlation(i, t)
}
\arguments{
\item{i}{i is a dichotomous-ordinal variables. x and y can also be the other way around.}

\item{t}{t is a continuous variables. x and y can also be the other way around.}
}
\description{
A biserial correlation is a correlation between dichotomous-ordinal and
continuous variables.
}
