% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ch02TestItemAnalysis.R
\name{CCRR}
\alias{CCRR}
\title{Conditional Correct Response Rate}
\usage{
CCRR(U, na = NULL, Z = NULL, w = NULL, ...)
}
\arguments{
\item{U}{U is a data matrix of the type matrix or data.frame.}

\item{na}{na argument specifies the numbers or characters to be treated as missing values.}

\item{Z}{Z is a missing indicator matrix of the type matrix or data.frame}

\item{w}{w is item weight vector}
}
\value{
A matrix of conditional correct response rates with Exametrika class.
Each element (i,j) represents the probability of correctly answering item j
given that item i was answered correctly.
}
\description{
The conditional correct response rate (CCRR) represents the ratio of the students
who passed Item C (consequent item) to those who passed Item A (antecedent item).
This function is applicable only to binary response data.
}
