% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BetweenItemAnalysis.R
\name{CCRR}
\alias{CCRR}
\title{Conditional Correct Repsonse Rate}
\usage{
CCRR(U, na = NULL, Z = NULL, w = NULL)
}
\arguments{
\item{U}{U is a data matrix of the type matrix or data.frame.}

\item{na}{na argument specifies the numbers or characters to be treated as missing values.}

\item{Z}{Z is a missing indicator matrix of the type matrix or data.frame}

\item{w}{w is item weight vector}
}
\description{
The conditional correct response rate(CCRR) represents the ratio of the studnets
who passed Item C(consequent item) to those who passed Item A(antecedent item)
}
