% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ch02TestItemAnalysis.R
\name{crr}
\alias{crr}
\title{Correct Response Rate}
\usage{
crr(U, na = NULL, Z = NULL, w = NULL)
}
\arguments{
\item{U}{U is a data matrix of the type matrix or data.frame.}

\item{na}{na argument specifies the numbers or characters to be treated as missing values.}

\item{Z}{Z is a missing indicator matrix of the type matrix or data.frame}

\item{w}{w is item weight vector}
}
\description{
The correct response rate (CRR) is one of the most basic and important
statistics for item analysis. This is an index of item difficulty and
a measure of how many students out of those who tried an item correctly
responded to it.
}
