% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudentAnalysis.R
\name{dataFormat}
\alias{dataFormat}
\title{dataFormat}
\usage{
dataFormat(U, na = NULL, Z = NULL, w = NULL)
}
\arguments{
\item{U}{is a data matrix of the type matrix or data.frame.}

\item{na}{na argument specifies the numbers or characters to be treated as missing values.}

\item{Z}{Z is a missing indicator matrix of the type matrix or data.frame}

\item{w}{w is item weight vector}
}
\description{
This function returns the sample size, data matrix, missing value identifier matrix,
and weight matrix from the given data matrix.
}
