% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudentAnalysis.R
\name{dataFormat}
\alias{dataFormat}
\title{dataFormat}
\usage{
dataFormat(data, na = NULL, id = 1, Z = NULL, w = NULL)
}
\arguments{
\item{data}{is a data matrix of the type matrix or data.frame.}

\item{na}{na argument specifies the numbers or characters to be treated as missing values.}

\item{id}{id indicates the column number containing the examinee ID. The default is 1.
If the answer pattern is contained in the first column, it is treated as if there is no ID vector.}

\item{Z}{Z is a missing indicator matrix of the type matrix or data.frame}

\item{w}{w is item weight vector}
}
\description{
dataFormat
}
