% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_Biclustering.R
\name{FieldAnalysis}
\alias{FieldAnalysis}
\title{Field Analysis}
\usage{
FieldAnalysis(x, digits = 4)
}
\arguments{
\item{x}{Biclustering Objects yielded by Biclustering Function}

\item{digits}{printed digits}
}
\value{
Returns a list of class c("exametrika", "Biclustering", "FieldAnalysis") containing:
\describe{
\item{FieldAnalysisMatrix}{A matrix showing field analysis results with rows
representing items and columns showing:
\describe{
\item{CRR}{Correct Response Rate}
\item{LFE}{Latent Field Estimation}
\item{Field1...FieldN}{Field membership values}
}
}
}
}
\description{
Output for Field Analysis
}
