% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch07Biclustering.R
\name{FieldAnalysis}
\alias{FieldAnalysis}
\title{Field Analysis}
\usage{
FieldAnalysis(x, digits = 4)
}
\arguments{
\item{x}{Biclustering Objects yielded by Biclustering Function}

\item{digits}{printed digits}
}
\description{
output for Field Analysis
}
