% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_grm.R
\name{grm_iif}
\alias{grm_iif}
\title{Item Infromation Function for grm}
\usage{
grm_iif(theta, a, b)
}
\arguments{
\item{theta}{latent score of subject}

\item{a}{discriminant parameter of IRF}

\item{b}{difficulty parameter of IRF}
}
\description{
Item Infromation Function for grm
}
