% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_GRM.R
\name{grm_prob}
\alias{grm_prob}
\title{Probability function for GRM}
\usage{
grm_prob(theta, a, b)
}
\arguments{
\item{theta}{Latent trait value of the subject}

\item{a}{Discrimination parameter of IRF}

\item{b}{Vector of difficulty parameters (thresholds) of IRF}
}
\value{
Vector of category selection probabilities
}
\description{
Calculates the probability of selecting each category given a latent trait value and item parameters.
}
\examples{
\dontrun{
# Example for an item with 3 categories
a <- 1.5
b <- c(-1.0, 1.0)
theta <- 0
grm_prob(theta, a, b)
}
}
