% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ch04IRTbaseFunctions.R
\name{IIF3PLM}
\alias{IIF3PLM}
\title{IIF for 3PLM}
\usage{
IIF3PLM(a, b, c, theta)
}
\arguments{
\item{a}{slope parameter}

\item{b}{locaiton parameter}

\item{c}{lower asymptote parameter}

\item{theta}{ability parameter}
}
\description{
Item Information Function for 3PLM
}
