% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleItemAnalysis.R
\name{ItemEntropy}
\alias{ItemEntropy}
\title{Item Entropy}
\usage{
ItemEntropy(U, na = NULL, Z = NULL, w = NULL)
}
\arguments{
\item{U}{U is a data matrix of the type matrix or data.frame.}

\item{na}{na argument specifies the numbers or characters to be treated as missing values.}

\item{Z}{Z is a missing indicator matrix of the type matrix or data.frame}

\item{w}{w is item weight vector}
}
\description{
The item entropy is an indicator of the variability or randomness
of the responses.
}
\details{
\eqn{e_j = -p_j\log_2p_j-(1-p_j)\log_2(1-p_j)}
}
