% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ch02TestItemAnalysis.R
\name{ItemLift}
\alias{ItemLift}
\title{Item Lift}
\usage{
ItemLift(U, na = NULL, Z = NULL, w = NULL)
}
\arguments{
\item{U}{U is a data matrix of the type matrix or data.frame.}

\item{na}{na argument specifies the numbers or characters to be treated as missing values.}

\item{Z}{Z is a missing indicator matrix of the type matrix or data.frame}

\item{w}{w is item weight vector}
}
\description{
The lift is a commonly used index in a POS data analysis.
The item lift of Item $k$ to Item $J$ is defined as follow:
\eqn{ l_{jk} = \frac{p_{k\mid j}}{p_k} }
}
\references{
Brin, S., Motwani, R., Ullman, J., & Tsur, S. (1997). Dynamic itemset counting and
implication rules for market basket data. In Proceedings of ACM SIGMOD International Conference
on Management of Data (pp. 255–264).https://dl.acm.org/doi/10.1145/253262.253325
}
