% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ch02TestItemAnalysis.R
\name{ItemOdds}
\alias{ItemOdds}
\title{Item Odds}
\usage{
ItemOdds(U, na = NULL, Z = NULL, w = NULL, ...)
}
\arguments{
\item{U}{U is a data matrix of the type matrix or data.frame.}

\item{na}{na argument specifies the numbers or characters to be treated as missing values.}

\item{Z}{Z is a missing indicator matrix of the type matrix or data.frame}

\item{w}{w is item weight vector}
}
\value{
A numeric vector of odds values for each item. Values range from 0 to infinity,
where:
\itemize{
\item odds > 1: correct response more likely than incorrect
\item odds = 1: equally likely
\item odds < 1: incorrect response more likely than correct
}
}
\description{
Item Odds are defined as the ratio of Correct Response Rate to
Incorrect Response Rate:
\deqn{O_j = \frac{p_j}{1-p_j}}
where \eqn{p_j} is the correct response rate for item j.
This function is applicable only to binary response data.

The odds value represents how many times more likely a correct response is
compared to an incorrect response. For example, an odds of 2 means students
are twice as likely to answer correctly as incorrectly.
}
\examples{
\dontrun{
# Easy item (80\% correct)
p1 <- 0.8
o1 <- p1/(1-p1)  # odds = 4

# Hard item (20\% correct)
p2 <- 0.2
o2 <- p2/(1-p2)  # odds = 0.25
}
}
