% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ch02SingleItemAnalysis.R
\name{ItemThreshold}
\alias{ItemThreshold}
\title{Item Threshold}
\usage{
ItemThreshold(U, na = NULL, Z = NULL, w = NULL)
}
\arguments{
\item{U}{U is a data matrix of the type matrix or data.frame.}

\item{na}{na argument specifies the numbers or characters to be treated as missing values.}

\item{Z}{Z is a missing indicator matrix of the type matrix or data.frame}

\item{w}{w is item weight vector}
}
\description{
Itemthreshold is a measure of difficulty based on a standard normal distribuiton.
}
