% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_explain.R
\docType{data}
\name{J12S5000}
\alias{J12S5000}
\title{J12S5000.Rdata}
\format{
A data frame with 5000 students and 12 items
}
\source{
\url{http://shojima.starfree.jp/tde/index.htm}
}
\usage{
J12S5000
}
\description{
J12S5000.Rdata
}
\keyword{datasets}
