% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_explain.R
\docType{data}
\name{J15S500}
\alias{J15S500}
\title{J15S500.Rdata}
\format{
A data frame with 500 students and 15 items
}
\source{
\url{http://shojima.starfree.jp/tde/index.htm}
}
\usage{
J15S500
}
\description{
J15S500.Rdata
}
\keyword{datasets}
