% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_explain.R
\docType{data}
\name{J20S400}
\alias{J20S400}
\title{J20S400.Rdata}
\format{
A data frame with 400 students and 20 items
}
\source{
\url{http://shojima.starfree.jp/tde/index.htm}
}
\usage{
J20S400
}
\description{
J20S400.Rdata
}
\keyword{datasets}
