% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_explain.R
\docType{data}
\name{J35S515}
\alias{J35S515}
\title{J35S515.Rdata}
\format{
A data frame with 515 students and 35 items
}
\source{
\url{http://shojima.starfree.jp/tde/index.htm}
}
\usage{
J35S515
}
\description{
J35S515.Rdata
}
\keyword{datasets}
