% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ch02InterItemAnalysis.R
\name{JointSampleSize}
\alias{JointSampleSize}
\title{Joint Sample Size}
\usage{
JointSampleSize(U, na = NULL, Z = NULL, w = NULL)
}
\arguments{
\item{U}{U is a data matrix of the type matrix or data.frame.}

\item{na}{na argument specifies the numbers or characters to be treated as missing values.}

\item{Z}{Z is a missing indicator matrix of the type matrix or data.frame}

\item{w}{w is item weight vector}
}
\description{
The joint sample size is a matrix whose elements are the number of
individuals who responded to each pair of items.
}
