% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch05LCA.R
\name{LCA}
\alias{LCA}
\title{Latent Clas Analysis}
\usage{
LCA(U, na = NULL, Z = NULL, w = NULL, ncls = 2)
}
\arguments{
\item{U}{U is a data matrix of the type matrix or data.frame.}

\item{na}{na argument specifies the numbers or characters to be treated as missing values.}

\item{Z}{Z is a missing indicator matrix of the type matrix or data.frame}

\item{w}{w is item weight vector}

\item{ncls}{number of latent class}
}
\description{
A function for estimating LCA using the EM algorithm.
}
