% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_LDLRA.R
\name{LD_param_est}
\alias{LD_param_est}
\title{LDparam set}
\usage{
LD_param_est(
  tmp,
  adj_list,
  classRefMat,
  ncls,
  smoothpost,
  beta1 = 2,
  beta2 = 2
)
}
\arguments{
\item{tmp}{tmp}

\item{adj_list}{adj_list}

\item{classRefMat}{values returned from emclus}

\item{ncls}{ncls}

\item{smoothpost}{smoothpost}

\item{beta1}{Beta distribution parameter 1 for prior density. Default is 2.}

\item{beta2}{Beta distribution parameter 2 for prior density. Default is 2.}
}
\description{
A function that extracts only the estimation of graph parameters
after the rank estimation is completed.
}
