% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch09LDLRA.R
\name{LD_param_est}
\alias{LD_param_est}
\title{LDparam set}
\usage{
LD_param_est(tmp, adj_list, classRefMat, ncls, smoothpost)
}
\arguments{
\item{tmp}{tmp}

\item{adj_list}{adj_list}

\item{classRefMat}{values returned from emclus}

\item{ncls}{ncls}

\item{smoothpost}{smoothpost}
}
\description{
A function that extracts only the estimation of graph parameters
after the rank estimation is completed.
}
