% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch09LDLRA.R
\name{LDLRA}
\alias{LDLRA}
\title{Local Dependence Latent Rank Analysis}
\usage{
LDLRA(
  U,
  Z = NULL,
  w = NULL,
  na = NULL,
  ncls = 2,
  method = "R",
  g_list = NULL,
  adj_list = NULL,
  adj_file = NULL,
  verbose = FALSE
)
}
\arguments{
\item{U}{U is either a data class of Exametrika, or raw data. When raw data is given,
it is converted to the Exametrika class with the \link{dataFormat} function.}

\item{Z}{Z is a missing indicator matrix of the type matrix or data.frame}

\item{w}{w is item weight vector}

\item{na}{na argument specifies the numbers or characters to be treated as missing values.}

\item{ncls}{number of latent class(rank). The default is 2.}

\item{method}{specify the model to analyze the data.Lcal dependence latent
class model is set to "C", latent rank model is set "R". The default is "R".}

\item{g_list}{A list compiling graph-type objects for each rank/class.}

\item{adj_list}{A list compiling matrix-type adjacency matrices for each rank/class.}

\item{adj_file}{A file detailing the relationships of the graph for each rank/class,
listed in the order of starting point, ending point, and rank(class).}

\item{verbose}{verbose output Flag. default is TRUE}
}
\value{
\describe{
\item{nobs}{Sample size. The number of rows in the dataset.}
\item{testlength}{Length of the test. The number of items included in the test.}
\item{crr}{correct response ratio}
\item{adj_list}{adjacency matrix list}
\item{g_list}{graph list}
\item{referenceMatrix}{Learned Parameters.A three-dimensional array of patterns where
item x rank x pattern.}
\item{IRP}{Marginal Item Reference Matrix}
\item{IRPIndex}{IRP Indices which include Alpha, Beta, Gamma.}
\item{TRP}{Test Reference Profile matrix.}
\item{LRD}{latent Rank/Class Distribution}
\item{RMD}{Rank/Class Membersip Distribution}
\item{TestFitIndices}{Overall fit index for the test.See also \link{TestFit}}
\item{Estimation_table}{Esitated parameters tables.}
\item{CCRR_table}{Correct Response Rate tables}
\item{Studens}{Student information. It includes estimated class
membership, probability of class membership, RUO, and RDO.}
}
}
\description{
performs local dependence latent lank analysis(LD_LRA) by Shojima(2011)
}
\details{
This function is intended to perform LD-LRA. LD-LRA is an analysis that
combines LRA and BNM, and it is used to analyze the network structure among
items in the latent rank. In this function, structural learning is not
performed, so you need to provide item graphs for each rank as separate files.
The file format for this is plain text CSV that includes edges (From, To) and
rank numbers.
}
