% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ch04IRTbaseFunctions.R
\name{LogisticModel}
\alias{LogisticModel}
\title{Four-Parameter Logistic Model}
\usage{
LogisticModel(a = 1, b, c = 0, d = 1, theta)
}
\arguments{
\item{a}{slope parameter}

\item{b}{locaiton parameter}

\item{c}{lower asymptote parameter}

\item{d}{upper asymptote parameter}

\item{theta}{ability parameter}
}
\description{
The four-parameter logistic model is a model where one additional
parameter d, called the upper asymptote parameter, is added to the
3PLM.
}
