% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch06LRA.R
\name{LRA}
\alias{LRA}
\title{Latent Rank Analysis}
\usage{
LRA(
  U,
  ncls = 2,
  na = NULL,
  Z = NULL,
  w = NULL,
  method = "GTM",
  mic = FALSE,
  maxiter = 100,
  BIC.check = FALSE
)
}
\arguments{
\item{U}{U is either a data class of Exametrika, or raw data. When raw data is given,
it is converted to the Exametrika class with the \link{dataFormat} function.}

\item{ncls}{number of latent class}

\item{na}{na argument specifies the numbers or characters to be treated as missing values.}

\item{Z}{Z is a missing indicator matrix of the type matrix or data.frame}

\item{w}{w is item weight vector}

\item{method}{Specify either "SOM" or "GTM".
SOM refers to the estimation method using Self-Organizing Mapping,
which is suitable when the data size is small. However, as the sample
size increases, it takes time to execute.
GTM is a batch learning type of SOM, equivalent to applying a gentle
filter to LCA (Shojima, 2022).}

\item{mic}{Monotonic increasing IRP option. The default is FALSE.}

\item{maxiter}{Maximum number of iterations.}

\item{BIC.check}{During estimation with SOM, this parameter determines
whether to use the change in BIC as the convergence criterion.
By default, it is FALSE and iteration continues until the maximum
number of iterations is reached. If set to TRUE, iteration
continues until the overall change in BIC falls below a negligible
amount, or until the iteration count reaches ten times the maximum
number of iterations.}
}
\value{
\describe{
\item{nobs}{Sample size. The number of rows in the dataset.}
\item{testlength}{Length of the test. The number of items included in the test.}
\item{Nclass}{number of classes you set}
\item{TRP}{Test Reference Profile matrix. The TRP is the column sum vector of estimated class reference matrix,
\eqn{\hat{\Pi}_c}}
\item{LCD}{Latent Class Dstribution table.see also \link{plot.Exametrika}}
\item{CMD}{Class Membership Dstribution table. see also \link{plot.Exametrika}}
\item{Students}{Class Membership Profile matrix.The s-th row vector of \eqn{\hat{M}_c}, \eqn{\hat{m}_c}, is the
class membership profile of Student s, namely the posterior probability distribution representing the student's
belonging to the respective latent classes. It also includes the rank with the maximum estimated membership probability,
as well as the rank-up odds and rank-down odds.}
\item{IRP}{Item Reference Profile matrix.The IRP of item j is the j-th row vector in the class reference matrix,
\eqn{\hat{\pi}_c}}
\item{IRPIndex}{The IRP information includes the item location parameters B and Beta,
the slope parameters A and Alpha, and the monotonicity indices C and Gamma.}
\item{ItemFitIndices}{Fit index for each item.See also \link{ModelFit}}
\item{TestFitIndices}{Overall fit index for the test.See also \link{ModelFit}}
}
}
\description{
A function for estimating LRA by SOM/GTM
}
