% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudentAnalysis.R
\name{nrs}
\alias{nrs}
\title{nrs}
\usage{
nrs(U, Z = NULL, w = NULL)
}
\arguments{
\item{U}{U is a data matrix of the type matrix or data.frame.}

\item{Z}{Z is a missing indicator matrix of the type matrix or data.frame}

\item{w}{w is item weight vector}
}
\value{
This function counts and returns the number of correct answers from the data matrix.
}
\description{
\code{nrs} The Number-right score (NRS) function returns the count of the items passed to it.
}
\examples{
# nrs(x)
}
