% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudentAnalysis.R
\name{percentile}
\alias{percentile}
\title{percentile}
\usage{
percentile(U, na = NULL, Z = NULL, w = NULL)
}
\arguments{
\item{U}{U is a data matrix of the type matrix or data.frame.}

\item{na}{na argument specifies the numbers or characters to be treated as missing values.}

\item{Z}{Z is a missing indicator matrix of the type matrix or data.frame}

\item{w}{w is item weight vector}
}
\value{
This function returns standardized score for each students.
}
\description{
The percentile function returns the corresponding score percentile,
out of 100 divisions, for each student.
}
\examples{
# percentile(U)
}
