% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExametrikaPlot.R
\name{plot.Exametrika}
\alias{plot.Exametrika}
\title{Plot Exametrika Output}
\usage{
\method{plot}{Exametrika}(
  x,
  type = c("IIC", "ICC", "TIC", "IRP", "TRP", "LCD"),
  items = NULL,
  students = NULL,
  ncol = 1,
  nrow = 1,
  ...
)
}
\arguments{
\item{x}{Exametrika Class object}

\item{type}{Plot type}

\item{items}{Plot item number}

\item{students}{Plot students ID}

\item{ncol}{Specifying the number of columns when there are many plots to be drawn}

\item{nrow}{Specifying the number of rows when there are many plots to be drawn}

\item{...}{othre optios}
}
\description{
Plot format for Exametria Class
}
