% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExametrikaPlot.R
\name{plot.Exametrika}
\alias{plot.Exametrika}
\title{Plot Exametrika Output}
\usage{
\method{plot}{Exametrika}(x, type = c("IIC", "ICC", "TIC"), items = NULL, ...)
}
\arguments{
\item{x}{Exametrika Class object}

\item{type}{Plot type}

\item{items}{Plot item number}

\item{...}{othre optios}
}
\description{
Plot format for Exametria Class
}
