% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ch00OutPutFunctions.R
\name{print.Exametrika}
\alias{print.Exametrika}
\title{print.exametrika}
\usage{
\method{print}{Exametrika}(x, digits = 3, ...)
}
\arguments{
\item{x}{Exametrika Class object}

\item{digits}{printed digits}

\item{...}{othre options}
}
\description{
Output format for Exametria Class
}
