% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ch04IRTbaseFunctions.R
\name{RaschModel}
\alias{RaschModel}
\title{Rasch Model}
\usage{
RaschModel(b, theta)
}
\arguments{
\item{b}{slope parameter}

\item{theta}{ability parameter}
}
\description{
The one-parameter logistic model is a model with only one parameter b.
This model is a 2PLM model in which a is constarind to 1.
This model is also called the Rasch model.
}
