% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudentAnalysis.R
\name{sscore}
\alias{sscore}
\title{sscore}
\usage{
sscore(U, Z = NULL, w = NULL)
}
\arguments{
\item{U}{U is a data matrix of the type matrix or data.frame.}

\item{Z}{Z is a missing indicator matrix of the type matrix or data.frame}

\item{w}{w is item weight vector}
}
\value{
This function returns standardized score for each students.
}
\description{
The standardized score indicates how high or low the student's ability is
placed in the standard normal distribution.
}
\examples{
# sscore(x)
}
