% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ch02StudentAnalysis.R
\name{stanine}
\alias{stanine}
\title{stanine}
\usage{
stanine(U, na = NULL, Z = NULL, w = NULL)
}
\arguments{
\item{U}{U is a data matrix of the type matrix or data.frame.}

\item{na}{na argument specifies the numbers or characters to be treated as missing values.}

\item{Z}{Z is a missing indicator matrix of the type matrix or data.frame}

\item{w}{w is item weight vector}
}
\value{
This function returns Stanine Rank for each students.
}
\description{
The Stanine scoring system divides students into nine groups.
These groups correspond to the following percentile ranges:
the lowest 4\%, the subsequent 7\%, the following 12\%, the next 17\%,
the middle 20\%, the subsequent 17\%, the following 12\%,
the next 7\%, and the highest 4\%.
}
\examples{
# stanine(U)
}
\references{
Angoff, W. H. (1984). Scales, norms, and equivalent scores. Educational Testing Service.
(Reprint of chapter in R. L. Thorndike (Ed.) (1971) Educational Measurement (2nd Ed.).
American Councilon Education.
}
