% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ch04IRTbaseFunctions.R
\name{ThreePLM}
\alias{ThreePLM}
\title{Three-Parameter Logistic Model}
\usage{
ThreePLM(b, theta)

ThreePLM(b, theta)

ThreePLM(b, theta)
}
\arguments{
\item{b}{slope parameter}

\item{theta}{ability parameter}

\item{a}{slope parameter}

\item{c}{lower asymptote parameter}

\item{d}{upper asymptote parameter}
}
\description{
The three-parameter logistic model is a model where the lower
asymptote parameter c is added to the 2PLM

The four-parameter logistic model is a model where one additional
parameter d, called the upper asymptote parameter, is added to the
3PLM.

The one-parameter logistic model is a model with only one parameter b.
This model is a 2PLM model in which a is constarind to 1.
This model is also called the Rasch model.
}
