% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ch04IRTbaseFunctions.R
\name{TwoPLM}
\alias{TwoPLM}
\title{Two-Parameter Logistic Model}
\usage{
TwoPLM(a, b, theta)
}
\arguments{
\item{a}{slope parameter}

\item{b}{locaiton parameter}

\item{theta}{ability parameter}
}
\description{
The two-parameter logistic model is a classic model that defines
the probability of a student with ability theta successfully
answering item j, using both a slope parameter and a
location parameter.
}
